# backend/static/openapi.yaml
openapi: 3.0.3

info:
  title: GNN Classifier API
  version: 0.1.1
  description: |
    Default system and infrastructure endpoints.
    Includes health checks, connectivity tests, and user management.

servers:
  - url: http://127.0.0.1:5000
    description: Local development server

paths:
  /users/create:
    post:
      summary: Create a user
      description: Creates a new user in PostgreSQL.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - full_name
                - username
                - email
                - password
                - avatar_color
              properties:
                full_name:
                  type: string
                  example: John Doe
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  example: secret123
                avatar_color:
                  type: string
                  example: "#00ffaa"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  username:
                    type: string
                    example: johndoe
                  email:
                    type: string
                    example: john@example.com

  /users/login:
    post:
      tags:
        - Users
        - Authentication
      summary: User login
      description: >
        Authenticate a user using username or email and password.
        Returns basic user information on success.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Username or email
                  example: johndoe@example.com
                password:
                  type: string
                  format: password
                  example: supersecretpassword
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  username:
                    type: string
                    example: johndoe
                  email:
                    type: string
                    example: johndoe@example.com
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid credentials
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing identifier or password
        "500":
          description: Internal server error

  /users/check:
    get:
      tags:
        - Users
      summary: Check username availability
      description: >
        Checks whether a given username is available for registration.
      operationId: checkUsernameAvailability
      parameters:
        - name: username
          in: query
          required: true
          description: Username to check for availability
          schema:
            type: string
            minLength: 1
            example: johndoe
      responses:
        "200":
          description: Username availability result
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    example: true
        "400":
          description: Missing or invalid username parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: username is required

  /users/test-create:
    get:
      summary: Create a test user
      description: Creates a hardcoded test user (development only).
      tags: [Users]
      responses:
        "200":
          description: Test user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 999
                  username:
                    type: string
                    example: route_user
                  email:
                    type: string
                    example: route@example.com

  /health:
    get:
      summary: Application health check
      tags: [Default]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /ping:
    get:
      summary: Database connectivity check
      tags: [Default]
      responses:
        "200":
          description: Database reachable
        "500":
          description: Database unreachable

  /teapot:
    get:
      summary: Easter egg
      tags: [Default]
      responses:
        "418":
          description: I'm a teapot ☕

tags:
  - name: Default
    description: Health and utility endpoints
  - name: Users
    description: User-related endpoints